requirements.txt
loguru
requests
pyyaml
jinja2
pydantic
cachetools
nltk
pillow
transformers